face-service/requirements.txt
fastapi
uvicorn
python-multipart
numpy
opencv-python-headless
gunicorn
transformers
torch
pillow
scipy
